Move QueryClient into subfolder
