+++ packages/stargate/src/queries/index.ts
@@ -1,1 +1,7 @@
+// Base symbols
+
+export { QueryClient } from "./queryclient";
+
+// Extensions
+
 export { BankExtension, setupBankExtension } from "./bank";
