+++ packages/stargate/types/queries/bank.d.ts
@@ -1,5 +1,5 @@
 import { cosmos } from "../generated/codecimpl";
-import { QueryClient } from "../queryclient";
+import { QueryClient } from "./queryclient";
 export interface BankExtension {
   readonly bank: {
     readonly balance: (address: string, denom: string) => Promise<cosmos.ICoin | null>;
