+++ packages/stargate/types/queries/index.d.ts
@@ -1,1 +1,2 @@
+export { QueryClient } from "./queryclient";
 export { BankExtension, setupBankExtension } from "./bank";
