+++ packages/stargate/types/queries/queryclient.d.ts
